Add check to not spam the sshd_config file
The `install.sh` script currently checks if there is a line that reads
`AuthorizedKeysCommand none` and then changes it. Old behaviour was to
simply append the line on the end of the file is this line didn't exist.

There is now a check to see if the correct line that we want already
exists, so we don't spam the end of the configuration file.