Add SELinux support to install script
If we have the `getenforce` command available to us, and we find that
it returns the value `Enforcing` (meaning SELinux will block our access
to AWS when called by `sshd`), we enable the `nis_enabled` boolean
persistently in SELinux to inform SELinux that it should expect outbound
access from sshd and other login programs (like PAM) to remote servers
to get user account information.

This allows `sshd` to call the script to get the user's public key from IAM.

I also cleaned up the `systemd` conditional in the script to use the
simpler way to check if commands exist.